+++ config/AppConfig/app.py
@@ -12,7 +12,7 @@ SHORT_NAME = "VP"
 DESCRIPTION = "VibraPilot browser automation desktop application by Vib Tools."
 TAGLINE = "Authorized browser automation desktop application."
 
-VERSION = "1.0.6.20"
+VERSION = "1.0.6.21"
 CREATED_DATE = "2026-08-07"
 RELEASE_DATE = "2026-08-10"
 UPDATED_DATE = "2026-08-10"